feat: #19 searchResult api 수정 완료, maplist 완성
