+++ src/screens/search/SearchScreen.tsx
@@ -12,85 +12,13 @@ import API from "../../services/API";
 import axios from "axios";
 
 type ResgisterScreenProp = StackNavigationProp<RootStackParamList, 'Search'>;
-interface M0 {
-  latitude: number;
-  longitude: number;
-}
 
 function SearchScreen() {
   const [search, setSearch] = useState('');
-  const [myLocation, setMyLocation] = useState<M0>({latitude: 0, longitude: 0});
   const navigation = useNavigation<ResgisterScreenProp>();
-  const [storeList, setStoreList] = useState([]);
-    const [myStoreList, setMyStoreList] = useState([{
-        store_id: 0,
-        store_name : '',
-        distance: 0,
-        waiting: 0,
-        is_waiting: true,
-        information : '',
-        latitude: 0,
-        longitude: 0
-    }]);
-
-  async function getFCMToken() {
-    // 가져오기
-    AsyncStorage.getItem('FCMToken', (err, FCMToken) => {
-        console.log("[FCMToken] " + FCMToken);
-        return(FCMToken);
-    });
-  }
-
-  useEffect(() => {
-    if (geolocation) {
-        geolocation.getCurrentPosition(success, error);
-        mapData();
-    }
-
-// 위치추적에 성공했을때 위치 값을 넣어줍니다.
-    function success(position: any) {
-        console.log (position.coords.latitude, position.coords.longitude)
-        setMyLocation({
-        latitude: position.coords.latitude,
-        longitude: position.coords.longitude,
-        });
-    }
-
-// 위치 추적에 실패 했을때 초기값을 넣어줍니다.
-    function error() { // 현우네 양꼬치 & 사천훠궈, 현우네 반점
-        setMyLocation({latitude: 37.5488, longitude: 127.1717 }); // 현우네밥집 주소 latitude: 37.5634, longitude: 126.9093 
-    }
-
-    async function mapData() {
-        try {
-            const response = await axios.post(
-                'http://15.164.28.246:8000/api/v1/stores/search/',
-                {
-                    token: await AsyncStorage.getItem('FCMToken', (err, res) => 
-                            {return(res);}),
-                    search: search,
-                    latitude: myLocation.latitude, 
-                    longitude: myLocation.longitude
-                },
-                { headers : {
-                  Authorization: await AsyncStorage.getItem('accessToken', (err, res) => 
-                                {return(res);})
-              }},)
-          .then(function (response) {
-            console.log(success)
-            setStoreList(response.data);
-            setMyStoreList(storeList["data"])
-          })
-          .catch(function (error) {
-            console.log(error)
-          });
-        } catch (error) {
-            console.log(error);
-        }
-        }}, []);
 
   const onPress = () => {{
-    navigation.navigate('SearchResult')
+    navigation.navigate('SearchResult', {search: search})
   }}
 
   return (
